spring email

+++ src/main/java/com/metrodatambkm/security/model/PrivilegeModel.java
@@ -11,12 +11,12 @@ import java.util.Set;
 
 @Entity
 @Table(name = "privilege")
-//@Data @AllArgsConstructor @NoArgsConstructor
+//@Data @AllArgsConstructor @NoArgsConstructor // kalau mau pakai ini, gaperlu bikin relasi ke entitas satunya
 public class PrivilegeModel implements Serializable {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "id")
-    private Integer id;
+    private Long id;
 
     @Column(name = "name", length = 50)
     private String name;
@@ -26,11 +26,11 @@ public class PrivilegeModel implements Serializable {
     @JsonIgnore
     private Set<RoleModel> privilegeRole;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
